implemented periodic boundary conditions for the gaugeconfig. note: the access w/ () now takes std::vector<long int> as argument for the coordinate x.

+++ su2code/hotstartImplementation.h
@@ -11,12 +11,12 @@ Gaugeconfig hotStart(const size_t timeSize,
 					 const double delta)
 {
 	auto output = Gaugeconfig(timeSize, spaceSize);
-	std::vector<size_t> x(4, 0);
+	std::vector<long int> x(4, 0);
 
-	for (x[0] = 0; x[0] < timeSize; x[0]++)
-		for (x[1] = 0; x[1] < spaceSize; x[1]++)
-			for (x[2] = 0; x[2] < spaceSize; x[2]++)
-				for (x[3] = 0; x[3] < spaceSize; x[3]++)
+	for (x[0] = 0; x[0] < static_cast<long int>(timeSize); x[0]++)
+		for (x[1] = 0; x[1] < static_cast<long int>(spaceSize); x[1]++)
+			for (x[2] = 0; x[2] < static_cast<long int>(spaceSize); x[2]++)
+				for (x[3] = 0; x[3] < static_cast<long int>(spaceSize); x[3]++)
 					for (size_t mu {0}; mu < Gaugeconfig::numSpacetimeDim; mu++)
 					{
 						output(x, mu) = randomSU2(engine, delta);
